RemarkableTexts/lib/requirements.txt
emoji==2.2.0
fpdf==1.7.2
imessage_reader==0.5.0
numpy==1.22.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.5.4.60
pdf2image==1.16.3
Pillow==9.1.0
Pillow==9.5.0
protobuf==4.23.0
thermalprinter==0.2.0
